refactor: 拆分context和util
