+++ src/components/form/formControl.tsx
@@ -15,7 +15,8 @@ import {
 } from "litten-hooks/dist/control/event/littenEvent.types";
 
 import { FormControlProps, FormItemValue } from "./form.types";
-import { FormContext } from "./form";
+import { FormContext } from "./context";
+
 
 export const FormControl = (props: FormControlProps) => {
     const { children, valuePath } = props;
